Added a system for storing the output
When there has been significant motion a frame will be stored in a video.mp4 file. Every 6 hours, that file will be closed, renamed with the start and end times, and moved to an external hard drive (if it is day). Motion detection now only uses 2 frames and returns the total number of contours, allowing for more accurate detection within the main thread. Also made sure that if the program is stopped everything saves gracefully.